requirements.txt
colorlog
openai-whisper
pyaudio
numpy
googletrans==3.1.0a0